requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
openai==1.35.3
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
python-dotenv==1.0.1
ruff==0.11.13
scipy==1.13.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
wsproto==1.2.0
